requirements.txt
httpx
beautifulsoup4
lxml
PyYAML
google-api-python-client
tqdm
